% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lctmc.R
\name{check_lctmc}
\alias{check_lctmc}
\title{Checks Model Specification of LCTMC}
\usage{
check_lctmc(
  data,
  X_names,
  W_names,
  K,
  par_constraint,
  controls,
  parallel_optim,
  MyModelName
)
}
\arguments{
\item{data}{reference \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}

\item{X_names}{reference \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}

\item{W_names}{reference \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}

\item{K}{reference \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}

\item{par_constraint}{reference \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}

\item{controls}{reference \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}

\item{parallel_optim}{reference \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}

\item{MyModelName}{reference \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}
}
\description{
To be used only within \code{lctmc_2x2()} or \code{lctmc_3x3()}. Takes the exact same arguments as
those functions and performs several checks to ensure the model is specified correctly.
}
\note{
This function checks the following conditions:
\enumerate{
\item \code{controls} is created from the \code{LCTMC::create_controls()} function
\item \code{parallel_optim} is a list object containing the correct element names
\item \code{MyModelName} is a character scalar, also checks for lengthy names
\item \code{scaling} parameter in \code{controls$fmt_data} matches \code{X_names} & \code{W_names}
\item \code{par_constraint} is either NULL or only zero constrain
\item \code{K} is an integer greater than 1
\item \code{controls$init02$which_step1} has the correct options
\item \code{data} contains a column called \code{id} as the individual identifying numbers
\item \code{data} contains \code{obsTime} & \code{state_at_obsTime}
\item \code{data} does not contain \code{x0} & \code{w0} as those are reserved for intercepts
\item \code{type} argument within \code{controls} has the correct options
\item \code{id} within \code{data} appears more than once (i.e., at least 2 obs or equivalently 1 transition)
}
}
\seealso{
\code{\link[=lctmc_2x2]{lctmc_2x2()}}; \code{\link[=lctmc_3x3]{lctmc_3x3()}}
}
