% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_P.R
\name{compute_P}
\alias{compute_P}
\alias{compute_P.lctmc_2x2}
\alias{compute_P.lctmc_3x3}
\title{Computes the transition probability matrix, \eqn{P}}
\usage{
compute_P(m, x1, x2, dt)

\method{compute_P}{lctmc_2x2}(m, x1, x2, dt)

\method{compute_P}{lctmc_3x3}(m, x1, x2, dt)
}
\arguments{
\item{m}{a 'lctmc_2x2' or 'lctmc_3x3' object obtained from the \code{lctmc} functions}

\item{x1}{a numeric scalar. It is the covariate corresponding to the parameters \code{beta1.12_k}, \code{beta1.21_k}, and \code{beta1.23_k}}

\item{x2}{a numeric scalar. It is the covariate corresponding to the parameters \code{beta2.12_k}, \code{beta2.21_k}, and \code{beta2.23_k}}

\item{dt}{a numeric scalar. This value indicates the time difference between observations.}
}
\value{
A list object of length equal to the number of latent classes specified when fitting the LCTMC model.
(See the documentation for \code{K} at \code{\link[=lctmc_2x2]{lctmc_2x2()}}). \cr
Each element of the list is the transition probability matrix for the corresponding latent class.
}
\description{
Uses the parameters estimated from the latent class CTMC model (\code{lctmc_2x2()} or \code{lctmc_3x3()}),
and computes the transition \strong{probability} matrices for some user specified covariate values
}
\note{
The transition \strong{probability} matrix is directly related to the transition \strong{rate} matrix, \eqn{Q}. \cr
The relation is the following:
\deqn{
  P = \exp(Q \cdot t)
}
where \eqn{Q} is obtained from the \code{compute_Q()} function, and \eqn{t} is equivalent to the input argument \code{dt}.
The exponential of a square matrix is defined as the Taylor expansion,
\deqn{
  \exp(Qt) = \sum_{n=0}^{\infty}\frac{1}{n!}(Qt)^{n}
}
}
\examples{
# load data
data("model_2x2", package = "LCTMC")
data("model_3x3", package = "LCTMC")


# Q matrix ---------
# 2x2
compute_Q(m = model_2x2, x1 = 1, x2 = 1)
# 3x3
compute_Q(m = model_3x3, x1 = 1, x2 = 1)


# Sojourn Times ---------
# 2x2
compute_sojourn(m = model_2x2, x1 = 1, x2 = 1)
# 3x3
compute_sojourn(m = model_3x3, x1 = 1, x2 = 1)


# Rate Ratios ---------
# 2x2
compute_RR(m = model_2x2, x1 = 1, x2 = 1)
# 3x3
compute_RR(m = model_3x3, x1 = 1, x2 = 1)


# Transition Probability ---------
# 2x2
compute_P(m = model_2x2, x1 = 1, x2 = 1, dt = 30)
# 3x3
compute_P(m = model_3x3, x1 = 1, x2 = 1, dt = 30)



# Transition Probability when t --> infty
lapply(
  X = compute_P(m = model_3x3, x1 = 1, x2 = 1, dt = 999999),
  FUN = function(x) round(x, 10)
)


# Transition Probability when t = 0
lapply(
  X = compute_P(m = model_3x3, x1 = 1, x2 = 1, dt = 0),
  FUN = function(x) round(x, 10)
)
}
\seealso{
\code{\link[=compute_Q]{compute_Q()}}; \code{\link[=get_P_2x2]{get_P_2x2()}}; \code{\link[=lctmc_2x2]{lctmc_2x2()}};
}
