% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_to_predict}
\alias{df_to_predict}
\title{Data sets to run predictions on}
\format{
A data frame with dimension \strong{52243 x 7}, columns have the following interpretation:
\describe{
\item{id}{ID number as the individual/person identifier (character)}
\item{obsTime}{time of observation/data collection (numeric)}
\item{state_at_obsTime}{the observed outcome corresponding to the time of observation/data collection}
\item{x1}{variable #1 that can affect the CTMC process (numeric)}
\item{x2}{variable #2 that can affect the CTMC process (numeric)}
\item{w1}{variable #1 that can affect the latent class probabilities (numeric)}
\item{w2}{variable #2 that can affect the latent class probabilities (numeric)}
\item{latent_class}{the underlying latent class/cluster identifier. \cr
When performing prediction on the latent class, this variable can be used to validate the result}
}
}
\usage{
data("df_to_predict")
}
\description{
A list object containing two data.frame objects: \verb{2x2} and \verb{3x3}. Each one has approximately 60 individuals. \cr
Among them are wo latent classes indicated by the \code{latent_class} column. Otherwise, other columns are structed the same
as the \code{example_df2x2} and \code{example_df3x3} data frame objects.
}
\keyword{datasets}
