% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_df2x2}
\alias{example_df2x2}
\title{A simulated data set for binary outcome.}
\format{
A data frame with dimension \strong{60895 x 7}, columns have the following interpretation:
\describe{
\item{id}{ID number as the individual/person identifier (character)}
\item{obsTime}{time of observation/data collection (numeric)}
\item{state_at_obsTime}{the observed outcome corresponding to the time of observation/data collection (numeric; 1 or 2)}
\item{x1}{variable #1 that can affect the CTMC process (numeric)}
\item{x2}{variable #2 that can affect the CTMC process (numeric)}
\item{w1}{variable #1 that can affect the latent class probabilities (numeric)}
\item{w2}{variable #2 that can affect the latent class probabilities (numeric)}
}
}
\usage{
data(example_df2x2)
}
\description{
The data set was simulated using the 'LCTMC.simulate' package. This data set
is meant to demonstrate how the input data should be formatted for a \strong{2x2} model,
and it can be used in conjunction with the example code under \code{\link[=lctmc_2x2]{lctmc_2x2()}}.
}
\keyword{datasets}
