% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_inits01_lctmc.R
\name{gen_inits01_lctmc}
\alias{gen_inits01_lctmc}
\alias{gen_inits01_lctmc_2x2}
\alias{gen_inits01_lctmc_3x3}
\title{Step 1/2 for generating initial value for model fitting}
\usage{
gen_inits01_lctmc_2x2(
  df,
  df_Xmat,
  df_Wmat,
  df_dt,
  K,
  par_constraint,
  N_sub,
  pct_keep,
  parallelize,
  parallel_optim
)

gen_inits01_lctmc_3x3(
  df,
  df_Xmat,
  df_Wmat,
  df_dt,
  K,
  par_constraint,
  N_sub,
  pct_keep,
  parallelize,
  parallel_optim
)
}
\arguments{
\item{df}{a data frame object containing row-wise transition data as binary variables. \cr
For example, if \code{trans.2_1} equals 1 then it means the observation was a transition from stage 2 to stage 1 within \code{df_dt} amount of time.}

\item{df_Xmat}{a matrix object housing the covariates that affect the CTMC portion of the model. \cr
This matrix should have the same number of rows as the data frame object, \code{df}}

\item{df_Wmat}{a matrix object housing the covariates that affect the latent classification part of the model. \cr
This matrix should have number of rows equal to unique number of individuals in the data frame object, \code{df}}

\item{df_dt}{a numeric vector housing the length of time interval between observations. This vector's length should be equal to number of rows in the data frame object, \code{df}}

\item{K}{an integer scalar. Use this variable to tell the function how many latent classes there should be. \cr}

\item{par_constraint}{See documentation in \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}

\item{N_sub}{See documentation in \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}

\item{pct_keep}{See documentation in \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}

\item{parallelize}{See documentation in \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}

\item{parallel_optim}{See documentation in \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}
}
\value{
A list containing the estimates obtained from the K-means algorithm.
This step outputs 3 elements:
\itemize{
\item \strong{step1} - a list of possible initial values, each element is the result of the K-means algorithm by trimming the individual estimate by the \code{pct_keep} argument
\item \strong{step1_full} - a single named numeric vector from the K-means using 100\% of individual estimates (no trimming outliers, when \code{pct_keep = 1})
\item \strong{step1_best} - a single named numeric vector that yielded the best observed log-likelihood value by testing trying each vector from \code{step1}
}
}
\description{
Performs step 1 of the initial value generation. This is a two-stage k-means procedure, where stage 1 involves fitting a
traditional CTMC model to each individual in the \code{df} data frame. Then the individual effects are pooled and fed into a k-means algorithm to split
the estimate into \code{K} clusters.
}
\note{
This is the second step out of six of fitting a latent class CTMC model (i.e., generate initial values via a two stage k-means algorithm). \cr\cr
This is also part 1 of initial value generation,
which is an individual level fitting algorithm and then uses K-means method to perform clustering based on the individuals' effects. \cr
If using This method alone to model the data, it will very likely perform poorly.
Especially when there are only few observation per person, it becomes reliable when each individual has many observations. \cr
Regardless, this approach will still produce usable initial values for the next steps.
}
\examples{
## TO BE ADDED ##
}
\seealso{
\code{\link[=lctmc_2x2]{lctmc_2x2()}}; \code{\link[=fmt_rowwise_trans]{fmt_rowwise_trans()}}; \code{\link[=gen_inits02_lctmc_2x2]{gen_inits02_lctmc_2x2()}}; \code{\link[=indiv_ctmc_2x2]{indiv_ctmc_2x2()}}
}
