% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_P.R
\name{get_P}
\alias{get_P}
\alias{get_P_2x2}
\alias{get_P_3x3}
\title{Computes transition probabilities}
\usage{
get_P_2x2(
  q12 = numeric(),
  q21 = numeric(),
  q23 = NA,
  log_q12 = numeric(),
  log_q21 = numeric(),
  log_q23 = NA,
  dt = numeric(),
  tol = 1e-15
)

get_P_3x3(
  q12 = numeric(),
  q21 = numeric(),
  q23 = numeric(),
  log_q12 = numeric(),
  log_q21 = numeric(),
  log_q23 = numeric(),
  dt = numeric(),
  tol = 1e-15
)
}
\arguments{
\item{q12}{a numeric vector for the transition rate from stage 1 to 2}

\item{q21}{a numeric vector for the transition rate from stage 2 to 1}

\item{q23}{a numeric vector for the transition rate from stage 2 to 3 (only applicable to 3x3 models)}

\item{log_q12}{a numeric vector for the \strong{log} transition rate from stage 1 to 2}

\item{log_q21}{a numeric vector for the \strong{log} transition rate from stage 2 to 1}

\item{log_q23}{a numeric vector for the \strong{log} transition rate from stage 2 to 3 (only applicable to 3x3 models)}

\item{dt}{a vector of numeric values for the time interval between observations}

\item{tol}{a numeric scalar. Tolerance for some quantities computed within the \code{get_P()} functions.
When the quantities are less than or equal to \code{tol}, then they are defaulted to \code{tol}}
}
\value{
a data.frame object with each column being one of the transition probability in the respective model. \cr\cr
For the 2x2 case, the possible transitions are \verb{1-1, 1-2, 2-1, 2-2},
and the corresponding transition probabilities are: \verb{P11, P12, P21, P22}. \cr
For the 3x3 case, the possible transitions are: \verb{1-1, 1-2, 1-3, 2-1, 2-2, 2-3}
and the corresponding transition probabilities are: \verb{P11, P12, P13, P21, P22, P23}.
In the case of exactly-observed data on \strong{state 3}, the transition densities are \verb{P13_exact, P23_exact}.
}
\description{
Computes the transition probabilities for the respective CTMC model (2x2 or 3x3). \cr
The only transition rates in this model are \eqn{q_{12}, q_{21} > 0}
}
\note{
The probabilities will be undefined when \eqn{q_{21}=0} and \eqn{q_{12} = q_{23} \geq 0}.
Thus, we've implemented a naive solution which applies the constrain \eqn{q_{21} \geq 10^{-12}}. \cr\cr
Additionally, the input argument should be of the same length, or their lengths should be multiples of one another.
Final note, this function does not need to depend on the number of latent classes because what we are computing is
the transition probability conditioned on the given latent class. \cr
}
\examples{
# 2x2 case ------------------
# transition rates
q12 = c(1, 0.5, 0.1)
q21 = c(1, 2, 3)

# time interval between observations
time_diffs = c(1, 5, 10)

# compute
LCTMC:::get_P_2x2(
  q12 = q12,
  q21 = q21,
  dt = time_diffs
)


# 3x3 case ------------------
# transition rates
q12 = c(1, 0.5, 0.1)
q21 = c(1, 2, 3)
q23 = c(1, 1, 1)

# time interval between observations
time_diffs = c(1, 5, 10)

# compute
LCTMC:::get_P_3x3(
  q12 = q12,
  q21 = q21,
  q23 = q23,
  dt = time_diffs
)
}
\seealso{
\code{\link[=Li_2x2]{Li_2x2()}}; \code{\link[=bik_all_2x2]{bik_all_2x2()}}
}
