% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_P_2x2.R
\name{get_P_2x2}
\alias{get_P_2x2}
\title{Computes transition probabilities (2x2)}
\usage{
get_P_2x2(q12 = c(), q21 = c(), dt = c())
}
\arguments{
\item{q12}{a numeric vector for the transition rate from stage 1 to 2}

\item{q21}{a numeric vector for the transition rate from stage 2 to 1}

\item{dt}{a numeric vector for the time difference between observations}
}
\value{
a data.frame object with 4 elements where each element is a resulting transition probability. \cr
For the 2x2 case, the possible transitions are \verb{1-1, 1-2, 2-1, 2-2} \cr
and the corresponding transition probabilities are: \verb{P11, P12, P21, P22}.
}
\description{
Computes the transition probabilities for the general 2x2 CTMC model. \cr
The only transition rates in this model are \eqn{q_{12}, q_{21} > 0}
}
\note{
The input argument should be of the same length, or their lengths should be multiples of one another. \cr
In addition, this function does not need to depend on the number of latent classes because what we are computing is
the transition probability conditioned on the given latent class.
}
\examples{
# transition rates
q12 = c(1, 0.5, 0.1)
q21 = c(1, 2, 3)

# time interval between observations
time_diffs = c(1, 5, 10)

# compute
LCTMC:::get_P_2x2(q12 = q12, q21 = q21, dt = time_diffs)
}
