% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_P_3x3.R
\name{get_P_3x3}
\alias{get_P_3x3}
\title{Computes transition probabilities (3x3)}
\usage{
get_P_3x3(q12 = c(), q21 = c(), q23 = c(), dt = c())
}
\arguments{
\item{q12}{a numeric vector for the transition rate from stage 1 to 2}

\item{q21}{a numeric vector for the transition rate from stage 2 to 1}

\item{q23}{a numeric vector for the transition rate from stage 2 to 3}

\item{dt}{a numeric vector for the time difference between observations}
}
\value{
a data.frame object with 8 (6+2) elements where each element is a transition probability/density. \cr
For this special 3x3 case, the possible transitions are: \verb{1-1, 1-2, 1-3, 2-1, 2-2, 2-3}
and the corresponding transition probabilities are: \verb{P11, P12, P13, P21, P22, P23}. \cr
In the case of exactly-observed data on \strong{state 3}, the transition densities are \verb{P13_exact, P23_exact}.
}
\description{
Computes the transition probabilities for a special 3x3 CTMC model. \cr
Specifically, the case with \eqn{q_{13} = q_{31} = q_{32} = q_{33} = 0}
}
\note{
The input argument should be of the same length, or their lengths should be multiples of one another. \cr
In addition, this function does not need to depend on the number of latent classes because what we are computing is
the transition probability conditioned on the given latent class. \cr
Additionally, when the absorbing state (Y=3) is exactly observed, such as in the case of deaths, the likelihood should use transition densities instead of probabilities.
}
\examples{
# transition rates
q12 = c(1, 0.5, 0.1)
q21 = c(1, 2, 3)
q23 = c(1, 1, 1)

# time interval between observations
time_diffs = c(1, 5, 10)

# compute
LCTMC:::get_P_3x3(q12 = q12, q21 = q21, q23 = q23, dt = time_diffs)
}
