% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SE_lctmc.R
\name{get_SE_lctmc}
\alias{get_SE_lctmc}
\alias{get_SE_lctmc_2x2}
\alias{get_SE_lctmc_3x3}
\title{SE Approximation for the Latent CTMC model}
\usage{
get_SE_lctmc_2x2(
  em,
  df,
  df.Xmat,
  df.Wmat,
  df.dt,
  K,
  par_constraint,
  solve.tol = (.Machine$double.eps)^2,
  symmetric.tol = 5e-11,
  eigen0.tol = 1e-10
)

get_SE_lctmc_3x3(
  em,
  df,
  df.Xmat,
  df.Wmat,
  df.dt,
  K,
  par_constraint,
  solve.tol = (.Machine$double.eps)^2,
  symmetric.tol = 5e-11,
  eigen0.tol = 1e-10
)
}
\arguments{
\item{em}{a list object of with the custom class 'lctmc_2x2.EM' or 'lctmc_3x3.EM'. This type of object is the output of the \code{EM_lctmc} functions}

\item{df}{a data frame object containing row-wise transition data as binary variables. \cr
For example, if \code{trans.2_1} equals 1 then it means the observation was a transition from stage 2 to stage 1 within \code{dt} amount of time.}

\item{df.Xmat}{a matrix object housing the covariates that affect the CTMC portion of the model. \cr
This matrix should have the same number of rows as the data frame object, \code{df}}

\item{df.Wmat}{a matrix object housing the covariates that affect the latent classification part of the model. \cr
This matrix should have number of rows equal to unique number of individuals in the data frame object, \code{df}}

\item{df.dt}{a numeric vector housing the length of time interval between observations. This vector's length should be equal to number of rows in the data frame object, \code{df}}

\item{K}{an integer scalar. Use this variable to tell the function how many latent classes there should be. \cr}

\item{par_constraint}{See documentation in \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}

\item{solve.tol}{a numeric scalar, typically a small decimal value. It is the tolerance for detecting linear dependencies in the hessian matrix. \cr
Defaults to \code{(.Machine$double.eps)^2} if not specified.}

\item{symmetric.tol}{a numeric scalar. Tolerane value for checking symmetric matrix. \cr Default is 5e-11}

\item{eigen0.tol}{a numeric scalar. Tolerance value for eigenvalues, any values smaller than this will be treated as 0. \cr Default is 1e-10}
}
\value{
a list object containing 4 elements:
\itemize{
\item \code{SE} is a data frame object containing columns for: the MLE, the approximated SE, and the 95\% confidence interval for the MLE
\item \code{covariance_code} a numeric scalar that can take be one of three values. \cr
0: when the covariance matrix is neither positive definite or semi positive definite \cr
1: when the covariance matrix is semi positive definite \cr
2: when the covariance matrix is positive definite
\item \code{hess_code} similar to \code{covariance_code} \cr
0: when the hessian is neither negative definite or semi negative definite (saddle point) \cr
1: when the covariance matrix is semi positive definite (inconclusive result) \cr
2: when the covariance matrix is positive definite (a local optimal point)
\item \code{Covariance} a matrix object which is the estimated variance covariance matrix for the estimated parameters. \cr
Note that the estimated SE is simply square root of the diagonal elements.
}
}
\description{
Uses the large sample approximation for to estimate the standard error of the MLE (see \strong{Note} for more info). \cr
The result of the approximation is the variance-covariance matrix for the vector of the estimated MLE,
where the diagonal elements are the variance and off-diagonals are the covariance.
}
\note{
This is step five out of six for fitting a latent class CTMC model (i.e., SE approximation via the hessian matrix). \cr\cr
This method relies on the large sample theory that
\deqn{
  \hat\theta \sim N(\theta, -H^{-1}(\hat\theta))
}
where, \eqn{H(\hat\theta)} is the hessian of the observed log likelihood function evaluated at the MLE. And \eqn{\hat\theta} is the estimated MLE.
Since this SE approximation relies on \eqn{N \rarr \infty}, user should be cautious when considering sample size and number of model parameters.
}
\examples{
## TO BE ADDED ##
}
\seealso{
\code{\link[=lctmc_2x2]{lctmc_2x2()}}; \code{\link[=EM_lctmc_2x2]{EM_lctmc_2x2()}}; \code{\link[=rescale_theta]{rescale_theta()}}
}
