% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_ctmc.R
\name{indiv_ctmc}
\alias{indiv_ctmc}
\alias{indiv_ctmc_2x2}
\alias{indiv_ctmc_3x3}
\title{Estimate the CTMC parameters for each individual}
\usage{
indiv_ctmc_2x2(subject_list, df, dt, Xmat, trace)

indiv_ctmc_3x3(subject_list, df, dt, Xmat, trace)
}
\arguments{
\item{subject_list}{a vector of ID numbers (charcter or numeric) matching with the \code{df$id}. The IDs in this vector should be unique}

\item{df}{a data frame object transformed via \code{fmt_rowwise_trans()}}

\item{dt}{a numeric vector holding the time interval between observations}

\item{Xmat}{a matrix of covariates that affects the CTMC process}

\item{trace}{a logical scalar. If TRUE then function will print progress update every 500 individuals fitted. If FALSE, no messages are printed.}
}
\value{
a matrix object housing the individual estimated CTMC parameter. \cr
The number of rows is equal to the number of individuals fitted. \cr
The number of columns is equal to 7, where the first six columns are the estimated coefficients and the last columns indicate whether convergence criteria is met for each person.
}
\description{
This function is part of the initial value generation process, where we first fit a CTMC model to each person in the input dataset.
Then the estimated effects are pooled and the K-means algorithm will be performed later to split the estimate into \code{K} clusters.
}
\examples{
## TO BE ADDED ##
}
