% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_2x2}
\alias{model_2x2}
\title{A 'lctmc' object}
\format{
A 'lctmc_2x2' object which behaves like a list object. See the documentation for \code{\link[=lctmc_2x2]{lctmc_2x2()}} for more info.
}
\usage{
data("model_2x2")
}
\description{
A 'lctmc_2x2' object obtained from the \code{\link[=lctmc_2x2]{lctmc_2x2()}} function. It is of type 'list'
and it is primarily used for running the example code for the various \code{compute} functions.
}
\seealso{
\code{\link[=lctmc_2x2]{lctmc_2x2()}}
}
\keyword{datasets}
