% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_3x3}
\alias{model_3x3}
\title{A 'lctmc' object}
\format{
A 'lctmc_3x3' object which behaves like a list object. See the documentation for \code{\link[=lctmc_3x3]{lctmc_3x3()}} for more info.
}
\usage{
data("model_3x3")
}
\description{
A 'lctmc_3x3' object obtained from the \code{\link[=lctmc_3x3]{lctmc_3x3()}} function. It is of type 'list'
and it is primarily used for running the example code for the various \code{compute} functions.
}
\seealso{
\code{\link[=lctmc_3x3]{lctmc_3x3()}}
}
\keyword{datasets}
