% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_2x2.R
\name{predict.lctmc_2x2}
\alias{predict.lctmc_2x2}
\title{Prediction function for the latent class CTMC model (2x2)}
\usage{
\method{predict}{lctmc_2x2}(object, ...)
}
\arguments{
\item{object}{an object of class "lctmc_2x2". Obtained by fitting the latent class CTMC model using \code{lctmc_2x2()}}

\item{...}{the following argument must be specified as part of \code{...}
\itemize{
\item \strong{df_pred} a data.frame object containing two columns: 'id' and 'obsTime'. \cr
The first is the identifier for each individual and the second is a numeric continuous variable, indicating the future time point that we wish to predict the disease status.
This time variable should be greater than the individual-wise maximum time specified in the \code{df_past} data frame
\item \strong{df_past} a data.frame object containing the history of observed disease status. This is used to perform latent clustering as well as predict disease state at a future time point
\item \strong{param.type} a character scalar. Either "mle" or "kmeans". \cr
If "mle" then the estimated MLE of the latent class CTMC is used as the parameter for prediction. If "kmeans" then Step 1 of the initial value generation process (K-Means algorithm) is used as the prediction parameters.
}}
}
\value{
A list object containing two elements:
\itemize{
\item \strong{df_pred.lc} is the predicted latent class classification. The predicted class is the class with the largest probability mass.
\item \strong{df_pred.ds} is the predicted disease status. This should be the same size as the input argument \code{df_pred} with an additional column added for the predicted class at the respective observation time.
}
}
\description{
Performs two types of prediction using the 2x2 latent CTMC model: \cr
\strong{(1)} predicts the latent cluster classification. \cr
\strong{(2)} predicts the disease status at a user specified time
}
\note{
In the \code{df_pred} data frame, there should only be one ID per person and "obsTime" must be a future time point relative to the \code{df_past} data frame.
The \code{df_past} data frame should contain the disease history of each person that is to be predicted. \cr
All covariates (x0, x1, x2, w0, w1, w2) should be included as well in the \code{df_past} data frame.
}
\examples{
## TO BE ADDED ##
}
