% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lctmc}
\alias{print.lctmc}
\title{Prints the fitted coefficients of a "lctmc" object}
\usage{
\method{print}{lctmc}(x, ...)
}
\arguments{
\item{x}{a "lctmc" object obtained from either the \code{lctmc_2x2()} or the \code{lctmc_3x3()} function}

\item{...}{NULL}
}
\description{
Prints estimated coefficient for a latent class CTMC model
}
\examples{
\dontrun{
  # first fit model via `lctmc_2x2()` or `lctmc_3x3()`
  my_model = lctmc_2x2(...)

  # prints
  print(x = my_model)

  # prints alternative
  my_model
}
}
