% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lctmc_control}
\alias{print.lctmc_control}
\title{Prints the LCTMC control object}
\usage{
\method{print}{lctmc_control}(x, ...)
}
\arguments{
\item{x}{a "lctmc_control" object obtained from either the \code{\link[=create_controls]{create_controls()}}}

\item{...}{NULL}
}
\description{
Prints the control elements of a "lctmc_control" object
}
\examples{
# load model
data('model_2x2', package = "LCTMC")


# prints
LCTMC:::print.lctmc(x = model_2x2)


# prints alternative
model_2x2
}
