% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sojourn.R
\name{compute_sojourn}
\alias{compute_sojourn}
\alias{compute_sojourn.lctmc_2x2}
\alias{compute_sojourn.lctmc_3x3}
\title{Computes the expected sojourn time}
\usage{
compute_sojourn(m, x1, x2)

\method{compute_sojourn}{lctmc_2x2}(m, x1, x2)

\method{compute_sojourn}{lctmc_3x3}(m, x1, x2)
}
\arguments{
\item{m}{a 'lctmc_2x2' or 'lctmc_3x3' object obtained from the \code{lctmc} functions}

\item{x1}{a numeric scalar. It is the covariate corresponding to the parameters \code{beta1.12_k}, \code{beta1.21_k}, and \code{beta1.23_k}}

\item{x2}{a numeric scalar. It is the covariate corresponding to the parameters \code{beta2.12_k}, \code{beta2.21_k}, and \code{beta2.23_k}}
}
\value{
A list object of length equal to the number of latent classes specified when fitting the LCTMC model.
(See the documentation for \code{K} at \code{\link[=lctmc_2x2]{lctmc_2x2()}}). \cr
Each element of the list is a named numeric vector of the expected sojourn time for the respective latent class
}
\description{
Uses the parameters estimated from the latent class CTMC model (\code{lctmc_2x2()} or \code{lctmc_3x3()}),
and computes the expected sojourn time for some user specified covariate values
}
\note{
The sojourn time is waiting time before a transition happens.
In CTMC models the sojourn is assumed to follow an exponential distribution. \cr
In the latent class CTMC model, the distribution is only exponential when we condition on the latent variable, \eqn{Z}
}
\examples{
# load data
data("model_2x2", package = "LCTMC")
data("model_3x3", package = "LCTMC")


# Q matrix ---------
# 2x2
compute_Q(m = model_2x2, x1 = 1, x2 = 1)
# 3x3
compute_Q(m = model_3x3, x1 = 1, x2 = 1)


# Sojourn Times ---------
# 2x2
compute_sojourn(m = model_2x2, x1 = 1, x2 = 1)
# 3x3
compute_sojourn(m = model_3x3, x1 = 1, x2 = 1)


# Rate Ratios ---------
# 2x2
compute_RR(m = model_2x2, x1 = 1, x2 = 1)
# 3x3
compute_RR(m = model_3x3, x1 = 1, x2 = 1)


# Transition Probability ---------
# 2x2
compute_P(m = model_2x2, x1 = 1, x2 = 1, dt = 30)
# 3x3
compute_P(m = model_3x3, x1 = 1, x2 = 1, dt = 30)



# Transition Probability when t --> infty
lapply(
  X = compute_P(m = model_3x3, x1 = 1, x2 = 1, dt = 999999),
  FUN = function(x) round(x, 10)
)


# Transition Probability when t = 0
lapply(
  X = compute_P(m = model_3x3, x1 = 1, x2 = 1, dt = 0),
  FUN = function(x) round(x, 10)
)
}
\seealso{
\code{\link[=compute_Q]{compute_Q()}}
}
