% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_controls.R
\name{create_controls}
\alias{create_controls}
\title{Creates the control object for fitting LCTMC}
\usage{
create_controls(type, ...)
}
\arguments{
\item{type}{a character scalar. Equals either "2x2" or "3x3" to indicate the number of category in the outcome variable.}

\item{...}{can be the following values:
\describe{
\item{scaling}{controls what factor covariate should be scaled to ensure algorithm stability. \cr
See \code{\link[=fmt_rowwise_trans]{fmt_rowwise_trans()}} for more info}
\item{trace}{controls whether data formatting function should print function details. \cr
See \code{\link[=fmt_rowwise_trans]{fmt_rowwise_trans()}} for more info}
\item{N_sub}{controls how many subjects to use for step 1 of initial value generation. \cr
See \code{\link[=gen_inits01_lctmc_2x2]{gen_inits01_lctmc_2x2()}} for more info}
\item{pct_keep}{controls how much of left/right tail should be truncated for the k-means algorithm. \cr
See \code{\link[=gen_inits01_lctmc_2x2]{gen_inits01_lctmc_2x2()}} for more info}
\item{parallelize.init01}{controls whether step 1 of initial value generation should be parallelized. \cr
See \code{\link[=gen_inits01_lctmc_2x2]{gen_inits01_lctmc_2x2()}} for more info}
\item{arg6}{...}
\item{arg7}{...}
\item{arg8}{...}
\item{arg9}{...}
\item{arg9}{...}
\item{arg9}{...}
}}
}
\value{
a custom class object that acts like a list. The output contains the following:
\describe{
\item{fmt_data}{contains elements: \code{scaling} and \code{trace}}
\item{init01}{...}
\item{init02}{...}
\item{EM}{...}
\item{SE}{...}
\item{rescale}{...}
\item{type}{controls whether contorls is for a 2x2 model or a 3x3 model.}
}
}
\description{
Create a custom object that acts like a list which serve as object that holds all control
parameters for fitting the LCTMC model.
}
\note{
XXXXXXXXXXXXXXXX
}
\examples{
# id column
my_ids = c(1, 1, 1,
           2, 2,
           3, 3, 3)

# observation times
c1 = c(0, 1, 1.5,
       0, 1,
       0, 2, 5)

# state at observation time
c2 = c(1, 2, 2,
       1, 1,
       2, 1, 3)

# some covariate data
x1 = c(rnorm(n = 3), rnorm(n = 2), rnorm(n = 3))
x2 = c(rep(1, 3), rep(0, 2), rep(1, 3))
w1 = c(rexp(n = 3, rate = 1), rexp(n = 2, rate = 1), rexp(n = 3, rate = 1))
w2 = c(rexp(n = 3, rate = 10), rexp(n = 2, rate = 10), rexp(n = 3, rate = 10))

# create data frame
my_df = data.frame(
  id = my_ids,
  obsTime = c1, state_at_obsTime = c2,
  x1 = x1, x2 = x2,
  w1 = w1, w2 = w2
)

# scaling parameters
my_scaling = c(x0 = 1, x1 = 0.01, x2 = 5, w0 = 1, w1 = 1000, w2 = 1, dt = 100)

# use function ~ 2x2 variation ~ note the 1-3 transition is not captured
LCTMC:::fmt_rowwise_trans(
  data = my_df,
  type = "2x2",
  X_names = c("x0", "x1", "x2"),
  W_names = c("w0", "w1", "w2"),
  scaling = my_scaling
)

# use function ~ 3x3 variation
LCTMC:::fmt_rowwise_trans(
  data = my_df,
  type = "3x3",
  X_names = c("x0", "x1", "x2"),
  W_names = c("w0", "w1", "w2"),
  scaling = my_scaling
)
}
\seealso{
\code{\link[=lctmc_2x2]{lctmc_2x2()}}; \code{\link[=lctmc_3x3]{lctmc_3x3()}}
}
