% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_rowwise_trans.R
\name{fmt_rowwise_trans}
\alias{fmt_rowwise_trans}
\title{Formats data for model fitting}
\usage{
fmt_rowwise_trans(
  data = data.frame(),
  type = c("2x2", "3x3"),
  X_names = c("x"),
  W_names = c("w"),
  scaling = c(x = 1, w = 1),
  ...
)
}
\arguments{
\item{data}{a data frame object. Data should be stored in long format. See example for data structure}

\item{type}{a character scalar. Equals either "2x2" or "3x3" to indicate the number of category in the outcome variable.}

\item{X_names}{a named character vector. Hosting the names of covariates for the CTMC model. It should be a column in \code{data}. \cr
Best set to x0, x1, x2 to avoid errors}

\item{W_names}{a named character vector. Hosting names of covariates for the latent class component. It should be a column in \code{data}. \cr
Best set to w0, w1, w2 to avoid errors}

\item{scaling}{a named numeric vector indicating which covariate are scaled by how much. Set to 1 for no scaling. \cr
For example, \code{scaling = c(x0 = 1, x1 = 0.01, x2 = 1, w0 = 1, w1 = 1, w2 = 1, dt = 0.002)}}

\item{...}{the following are optional parameters\
\describe{
\item{trace}{a logical scalar, if TRUE, function will print which parameters were scaled. Default is FALSE which does not print messages}
}}
}
\value{
a list object containing the 4 major data objects needed to perform model fitting. The objects are the following:
\describe{
\item{Xmat}{a numeric matrix of the "X" covariates which are variables that affect the CTMC process}
\item{Wmat}{a numeric matrix of the "W" covariates which are variables that affect the latent class probabilities}
\item{dt}{a numeric vector of the time interval between observations}
\item{df_trans}{a data.frame object containing the binary transition indicator variables}
}
}
\description{
Takes a data frame object in long format and convert into a row-wise transition format.
Input data should meet certain requirements indicated in the \strong{Note} section
}
\note{
This is the first step out of six of fitting a latent class CTMC model (i.e., data preparation). \cr\cr
Four conditions of the input data frame, \code{data}, are checked. If any fails, then error will be returned
\enumerate{
\item there must be a column named "id" in the input data
\item there must be columns named "state_at_obsTime" and "obsTime" in the input data
\item input data cannot have columns named "x0" and "w0" as these are reserved for the function to create a intercept term variable
\item each person indicated by the "id" column should have at least 2 observation minimum
}
}
\examples{
# id column
my_ids = c(1, 1, 1,
           2, 2,
           3, 3, 3)

# observation times
c1 = c(0, 1, 1.5,
       0, 1,
       0, 2, 5)

# state at observation time
c2 = c(1, 2, 2,
       1, 1,
       2, 1, 3)

# some covariate data
x1 = c(rnorm(n = 3), rnorm(n = 2), rnorm(n = 3))
x2 = c(rep(1, 3), rep(0, 2), rep(1, 3))
w1 = c(rexp(n = 3, rate = 1), rexp(n = 2, rate = 1), rexp(n = 3, rate = 1))
w2 = c(rexp(n = 3, rate = 10), rexp(n = 2, rate = 10), rexp(n = 3, rate = 10))

# create data frame
my_df = data.frame(
  id = my_ids,
  obsTime = c1, state_at_obsTime = c2,
  x1 = x1, x2 = x2,
  w1 = w1, w2 = w2
)

# scaling parameters
my_scaling = c(x0 = 1, x1 = 0.01, x2 = 5, w0 = 1, w1 = 1000, w2 = 1, dt = 100)

# use function ~ 2x2 variation ~ note the 1-3 transition is not captured
LCTMC:::fmt_rowwise_trans(
  data = my_df,
  type = "2x2",
  X_names = c("x0", "x1", "x2"),
  W_names = c("w0", "w1", "w2"),
  scaling = my_scaling
)

# use function ~ 3x3 variation
LCTMC:::fmt_rowwise_trans(
  data = my_df,
  type = "3x3",
  X_names = c("x0", "x1", "x2"),
  W_names = c("w0", "w1", "w2"),
  scaling = my_scaling
)
}
\seealso{
\code{\link[=lctmc_2x2]{lctmc_2x2()}}, \code{\link[=lctmc_3x3]{lctmc_3x3()}}
}
