% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_inits01_lctmc.R
\name{gen_inits01_lctmc}
\alias{gen_inits01_lctmc}
\alias{gen_inits01_lctmc_2x2}
\alias{gen_inits01_lctmc_3x3}
\title{Step 1/2 for generating initial value for model fitting}
\usage{
gen_inits01_lctmc_2x2(
  df,
  Xmat,
  Wmat,
  dt,
  K,
  N_sub,
  pct_keep,
  par_constraint,
  parallelize,
  parallel_optim
)

gen_inits01_lctmc_3x3(
  df,
  Xmat,
  Wmat,
  dt,
  K,
  N_sub,
  pct_keep,
  par_constraint,
  parallelize,
  parallel_optim
)
}
\arguments{
\item{df}{a data frame object containing row-wise transition data as binary variables. \cr
For example, if \code{trans.2_1} equals 1 then it means the observation was a transition from stage 2 to stage 1 within \code{dt} amount of time.}

\item{Xmat}{a matrix object housing the covariates that affect the CTMC portion of the model. \cr
This matrix should have the same number of rows as the data frame object, \code{data}}

\item{Wmat}{a matrix object housing the covariates that affect the latent classification part of the model. \cr
This matrix should have number of rows equal to unique number of individuals in the data frame object, \code{data}}

\item{dt}{a numeric vector housing the length of time interval between observations. This vector's length should be equal to number of rows in the data frame object, \code{data}}

\item{K}{the number of categories the latent class variable has}

\item{N_sub}{a numeric scalar. This is used for Step 1 of initial value generation where the algorithm fits the traditional CTMC model for each individual. \cr
Fitting the model for \emph{all} individuals might have long run time without improvement in the accuracy of the estimation. Thus, setting a maximum number of individuals to use for the initial value generation could shorten run without sacrificing in estimation accuracy.}

\item{pct_keep}{a numeric vector where each element of the vector ranges from 0 to 1 (ideally at minimum 0.50). \cr
This argument controls what percentage of the individual level estimated effects to be used for the K-means algorithm for initial value generation.
For example, for pct_keep = c(0.8), after individuals effects are estimated, only the 10th to 90th percentile are used for the K-Means algorithm to obtain cluster level estimates.}

\item{par_constraint}{a named numeric vector to indicate which parameter is constrained. Set equal to NULL for unconstrained model. \cr
For example, \code{c(alpha1.1 = 0)} constraints the parameter 'alpha1.1' to be a constant 0. \strong{NOTE:} Current version of the code will \emph{only} work with constrains equal to 0.}

\item{parallelize}{a logical scalar. Set to TRUE if we want the for-loop for the individual-wise CTMC to be parallelized}

\item{parallel_optim}{a list object telling the function whether parallel process should be used. \cr
The list should contain \strong{two} elements: \cr
(1) \code{run} a logical scalar, if TRUE then this function will use parallel processing. If FALSE, then the \code{cl} argument is ignored. \cr
(2) \code{cl} is an object obtained from the \code{parallel} package, for example \cr \code{cl = parallel::makeCluster(spec = 2)}}
}
\value{
A list containing the estimates obtained from the K-means algorithm.
This step outputs 3 elements:
\itemize{
\item \strong{step1} - a list of possible initial values, each element is the result of the K-means algorithm by trimming the individual estimate by the \code{pct_keep} argument
\item \strong{step1_full} - a single named numeric vector from the K-means using 100\% of individual estimates (no trimming outliers, when \code{pct_keep = 1})
\item \strong{step1_best} - a single named numeric vector that yielded the best observed log-likelihood value by testing trying each vector from \code{step1}
}
}
\description{
Performs step 1 of the initial value generation. This is a two-stage k-means procedure, where stage 1 involves fitting a
traditional CTMC model to each individual in the \code{df} data frame. Then the individual effects are pooled and fed into a k-means algorithm to split
the estimate into \code{K} clusters.
}
\note{
This is the second step out of six of fitting a latent class CTMC model (i.e., generate initial values via a two stage k-means algorithm). \cr\cr
This is also part 1 of initial value generation,
which is an individual level fitting algorithm and then uses K-means method to perform clustering based on the individuals' effects. \cr
If using This method alone to model the data, it will very likely perform poorly.
Especially when there are only few observation per person, it becomes reliable when each individual has many observations. \cr
Regardless, this approach will still produce usable initial values for the next steps.
}
\examples{
## TO BE ADDED ##
}
