% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_inits02_lctmc.R
\name{gen_inits02_lctmc}
\alias{gen_inits02_lctmc}
\alias{gen_inits02_lctmc_2x2}
\alias{gen_inits02_lctmc_3x3}
\title{Step 2/2 for generating initial value for model fitting}
\usage{
gen_inits02_lctmc_2x2(
  step2_inits,
  df,
  Xmat,
  Wmat,
  dt,
  K,
  par_constraint,
  parallel_optim
)

gen_inits02_lctmc_3x3(
  step2_inits,
  df,
  Xmat,
  Wmat,
  dt,
  K,
  par_constraint,
  parallel_optim
)
}
\arguments{
\item{step2_inits}{a named numeric vector. This should be a vector of model parameters. \cr
The names of the vector should be parameters names and the values are the initial value for direct likelihood optimization}

\item{df}{a data frame object containing row-wise transition data as binary variables. \cr
For example, if \code{trans.2_1} equals 1 then it means the observation was a transition from stage 2 to stage 1 within \code{dt} amount of time.}

\item{Xmat}{a matrix object housing the covariates that affect the CTMC portion of the model. \cr
This matrix should have the same number of rows as the data frame object, \code{df}}

\item{Wmat}{a matrix object housing the covariates that affect the latent classification part of the model. \cr
This matrix should have number of rows equal to unique number of individuals in the data frame object, \code{df}}

\item{dt}{a numeric vector housing the length of time interval between observations. This vector's length should be equal to number of rows in the data frame object, \code{df}}

\item{K}{an integer scalar. Use this variable to tell the function how many latent classes there should be. \cr}

\item{par_constraint}{See documentation in \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}

\item{parallel_optim}{See documentation in \code{\link[=lctmc_2x2]{lctmc_2x2()}} or \code{\link[=lctmc_3x3]{lctmc_3x3()}}}
}
\value{
a named numeric vector. This is the vector of model parameters that maximizes the observed data log-likelihood function (see \strong{Note} section).
}
\description{
Performs step 2 of the initial value generation. This is a direct optimization on the observed data likelihood function.
We condition on the number of latent clusters and sum over joint probability between the observed data and the unobserved data. \cr
That is,
\deqn{
  P(Y_{i}) = \sum_{k=1}^{K} P(Y_{i} \cap Z_{i}=k)
}
See \strong{Note} section for more info.
}
\note{
This is the third step out of six of fitting a latent class CTMC model (i.e., generate initial values via direct optimization). \cr\cr
This step is equivalent to obtaining the MLE. It performs numerical optimization on the observed data log likelihood function:
\deqn{
  log(P(Y)) = \sum_{i} log(P(Y_{i}))
}
where
\deqn{
  P(Y_{i}) = \sum_{k} P(Y_{i}=y_{i} \cap Z_{i}=k)
}
and the summand could be simplified as:
\deqn{
  P(Y_{i} \cap Z_{i}) = P(Z_{i}) \cdot P(Y_{i} | Z_{i})
}
}
\examples{
## TO BE ADDED ##
}
\seealso{
\code{\link[=lctmc_2x2]{lctmc_2x2()}}; \code{\link[=gen_inits01_lctmc_2x2]{gen_inits01_lctmc_2x2()}}; \code{\link[=EM_lctmc_2x2]{EM_lctmc_2x2()}}
}
