% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict.lctmc_2x2}
\alias{predict.lctmc_3x3}
\title{Prediction function for the latent class CTMC model}
\usage{
\method{predict}{lctmc_2x2}(object, ...)

\method{predict}{lctmc_3x3}(object, ...)
}
\arguments{
\item{object}{an object of class "lctmc_2x2" or "lctmc_3x3". Obtained by fitting the latent class CTMC model using the \code{lctmc} functions}

\item{...}{the following argument must be specified as part of \code{...}
\describe{
\item{df_past}{a data.frame object containing the history of observed disease status. \cr
This is used to perform latent clustering as well as predict disease state at a future time point}
\item{pred_threshold}{a numeric value between 0 and 1. By specifying this threshold, the predicted class will be assigned as the class with probability \cr
greater than this threshold \strong{and} has the largest probability mass. Defaults to 0.}
}}
}
\value{
A data.frame object containing the predicted latent class.
}
\description{
Input an "lctmc" model object and individuals' disease history. This function will predict the individuals' latent class.
}
\note{
The \code{df_past} data frame should contain the disease history of each person that is to be predicted. \cr
All covariates (x0, x1, x2, w0, w1, w2) should be included as well in the \code{df_past} data frame.
}
\examples{
# load data
data("df_to_predict", package = "LCTMC")
data("model_2x2", package = "LCTMC")

# create data sets (full data)
my_df_past.full = df_to_predict$`2x2`


# create data sets (truncated, 3 obs per person)
my_df_past.only3 = lapply(
  X = split(x = my_df_past.full, f = my_df_past.full$id),
  FUN = function(d) {
    d = d[1:3, ]
    d
  }
)
my_df_past.only3 = do.call(`rbind`, my_df_past.only3)


# create data sets (truncated, 6 obs per person)
my_df_past.only6 = lapply(
  X = split(x = my_df_past.full, f = my_df_past.full$id),
  FUN = function(d) {
    d = d[1:6, ]
    d
  }
)
my_df_past.only6 = do.call(`rbind`, my_df_past.only6)


# predict (1) ~ only 3 obs per person in disease history
my_predictions = predict(
  object = model_2x2,
  df_past = my_df_past.only3,
  pred_threshold = 0.7
)
table(my_predictions$pred_class, my_predictions$pred_class_strict)


# predict (2) ~ only 6 obs per person in disease history
my_predictions = predict(
  object = model_2x2,
  df_past = my_df_past.only6,
  pred_threshold = 0.7
)
table(my_predictions$pred_class, my_predictions$pred_class_strict)


# predict (3) ~ max obs per person in disease history
my_predictions = predict(
  object = model_2x2,
  df_past = my_df_past.full,
  pred_threshold = 0.7
)
table(my_predictions$pred_class, my_predictions$pred_class_strict)


# predict (4) ~ extreme threshold criteria, leading to all NA predictions
my_predictions = predict(
  object = model_2x2,
  df_past = my_df_past.only6,
  pred_threshold = 0.9999
)
table(my_predictions$pred_class, my_predictions$pred_class_strict)
}
\seealso{
\code{\link[=lctmc_2x2]{lctmc_2x2()}}; \code{\link[=lctmc_3x3]{lctmc_3x3()}}
}
