% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_theta.R
\name{rescale_theta}
\alias{rescale_theta}
\title{Re-scales model parameters}
\usage{
rescale_theta(df_theta = NULL, v_theta = NULL, scaling = numeric(), ...)
}
\arguments{
\item{df_theta}{a data frame object with data stored in long-format}

\item{v_theta}{a named numeric scalar. It is a scaling parameter for observation time intervals. \cr}

\item{scaling}{a named numeric vector indicating the factor of scaling. \cr
For example (1), if \code{scaling = c(x1 = 0.5)}, then the coefficient associated with 'x1' are multiplied by a factor of 0.5. \cr
For example (2), if \code{scaling = c(dt = 0.5)}, then the coefficient associated with 'dt' are added by a factor of log(0.5).}

\item{...}{two optional argument can be supplied:
\itemize{
\item \strong{mult_vars} a character vector, indicating which columns in \code{df_theta} needs multiplicative re-scaling
\item \strong{add_vars} a character vector, indicating which columns in \code{df_theta} needs additive re-scaling
}}
}
\value{
A list object containing the 2 elements:
\itemize{
\item \strong{df_theta} a data.frame object with its columns re-scaled accordingly. Its dimension is equal to the input \code{df_theta}. \cr
If the input \code{df_theta} is NULL then the output will also be NULL
\item \strong{v_theta} a named numeric vector with its elements re-scaled accordingly. Its length is equal to the input \code{v_theta}. \cr
If the input \code{v_theta} is NULL then the output will also be NULL
}
}
\description{
This function scales the model parameter by some user specified scaling factors.
It can take either parameters in data.frame format or in vector form.
}
\note{
This is the final step of fitting a latent class CTMC model (i.e., scaling estimated parameters back to the original data's units).
}
\examples{
### TO BE ADDED ###
}
\seealso{
\code{\link[=lctmc_2x2]{lctmc_2x2()}} \code{\link[=get_SE_lctmc_2x2]{get_SE_lctmc_2x2()}}
}
