% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_global_optim.R
\name{test_global_optim}
\alias{test_global_optim}
\alias{test_global_optim.lctmc_2x2}
\alias{test_global_optim.lctmc_3x3}
\title{Compare log likelihood between MLE and true parameters}
\usage{
test_global_optim(m, true_param, tol, data)

\method{test_global_optim}{lctmc_2x2}(m, true_param, tol, data)

\method{test_global_optim}{lctmc_3x3}(m, true_param, tol, data)
}
\arguments{
\item{m}{a 'lctmc_2x2' or 'lctmc_3x3' object obtained from the \code{lctmc} functions}

\item{true_param}{a list object housing the true parameter values. This object can be obtained from
the 'LCTMC.simulate' package via the \code{gen_true_param()} function.}

\item{tol}{a numeric scalar. Tolerance value when comparing the log-likelihood values.
The smaller this value, the more precise the comparison.}

\item{data}{a data.frame object. This should be the data set that the LCTMC model fitted on.}
}
\value{
A list object of length 3. The elements of this list are:
\describe{
\item{global_optim}{a logical scalar. TRUE if the \eqn{\log(L_{mle}) - \log(L_{true}) \geq \epsilon},
where \eqn{\epsilon} is the \code{tol} input argument.}
\item{L_mle}{a numeric scalar. The observed log-likelihood value when evaluated at the MLE.}
\item{L_true}{a numeric scalar. The observed log-likelihood value when evaluated at the true parameter values.}
}
}
\description{
\strong{SIMULATION STUDY ONLY} \cr
Input a fitted LCTMC model and provide some true parameter object (obtained from the 'LCTMC.simulate' package).
This function will then compute the log observed data log-likelihood evaluated at the MLE and the true parameter value.
If the MLE is truly a maximum likelihood estimate, then it should yield a larger log-likelihood value.
}
\note{
This function is specifically meant to be used during a simulation study. \cr
Currently, this function only supports when \code{K = 3} as the 'LCTMC.simulate' package only
simulates the case of 3 latent classes.
}
\seealso{
\code{\link[=lctmc_2x2]{lctmc_2x2()}}; \code{\link[=lctmc_3x3]{lctmc_3x3()}}
}
